Add calibration standards

Convenience function to add calibration standards to a continuous flow peak table (standalone or inside an isofiles object).

iso_add_standards(...)

# S3 method for iso_file_list
iso_add_standards(
  iso_files,
  stds,
  match_by = default(std_match_by),
  is_std_peak = default(is_std_peak),
  quiet = default(quiet)
)

# S3 method for data.frame
iso_add_standards(
  dt,
  stds,
  match_by = default(std_match_by),
  is_std_peak = default(is_std_peak),
  quiet = default(quiet),
  is_standard = NULL
)

Arguments

...
S3 method placeholder parameters, see class specific functions for details on parameters

iso_files
collection of continuous flow iso_file objects

stds
standards data frame

match_by
what column(s) to match the standards by

is_std_peak
new column that holds information about which ones are standard peaks (i.e. have known isotopic values)

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.

dt
data frame with the peak table

is_standard
renamed to is_std_peak because the naming caused too much confusion, will be removed in future versions, please use is_std_peak instead

Value

iso files with standards data frame merged into the peak table and the following information column added:

is_std_peak: a logical TRUE/FALSE indicating which data table entry is a standard

See also