Calculate deltas

Calculate delta values from recorded data in dual inlet files (continuous flow not yet supported). Note that these are recorded delta values of sample vs. standard in each dual inlet file and NOT normalized against any external standard or reported values of the standard. They usually need at least a frame shift correction afterwards. Also note that the delta values are only multipled by 1000 if in_permil=TRUE (the default), otherwise returns raw values without unit multiplication.

iso_calculate_deltas(...)

# S3 method for iso_file_list
iso_calculate_deltas(
  iso_files,
  deltas,
  bracket = TRUE,
  in_permil = TRUE,
  quiet = default(quiet)
)

# S3 method for data.frame
iso_calculate_deltas(
  df,
  deltas,
  bracket = TRUE,
  in_permil = TRUE,
  ratio_prefix = "r",
  quiet = default(quiet)
)

Arguments

...
S3 method placeholder parameters, see class specific functions for details on parameters

iso_files
collection of dual inlet iso_file objects

deltas
which deltas to calculate from the raw data, must be in the complete format with prefix 'd' and both numerator and denominator mass, e.g. c("d45/44", "d46/44") to calculate both delta 45/44 and delta 46/44. Deltas can only be calculated if the corresponding ratio columns already exist.

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.

df
a data frame of raw dual inlet data, must have at minimum the columns 'file_id', 'type' (with values "standard" and "sample") and 'cycle' (integer column), as well as all of the required ratio columns with the correct ratio_prefix.

ratio_prefix
the prefix of the ratio columns. Default is "r" which is what iso_calculate_ratios generates.

Value

the passed in data with deltas added

Details

If bracket=TRUE (the default), the delta values are calculated from the raw data ratios of the sample and the two bracketing standards in the same cycle and cycle before it (this implicitly assumes that the pre-standard has cycle number 0) If bracket=FALSE, the delta values are calculated straight from the corresponding raw data ratios of the sample and the standard in the same cycle. Delta values are only stored in the sample rows of the raw data, not in the standard rows.