Export calibration to Excel

This function exports the passed in calibration(s) data frame to Excel. The different kinds of information (all data, model coefficients, summary, evaluation range) are exported to separate tabs within the excel file. If there are multiple different calibrations in the data frame (i.e. multiple *calib_params columns), ALL will be exported with different tabs for each.

iso_export_calibration_to_excel(
  calibs_df,
  filepath,
  ...,
  include_all_data = TRUE,
  include_calib_coefs = TRUE,
  include_calib_summary = TRUE,
  include_calib_range = TRUE,
  with_explicit_units = TRUE,
  quiet = default(quiet)
)

Arguments

calibs_df
data frame that includes calibration columns (i.e. the output of functions like iso_generate_calibration, iso_apply_calibration, iso_evaluate_calibration_range). Can be the output from iso_get_calibration_data if called after a calibration has been generated.

filepath
the path where to store the Excel file.

...
named list of additional data frames to include in the summary (each will get their own tab)

include_all_data
whether to include a tab for all the peak data - if this function is called after iso_apply_calibration this will include

include_calib_coefs
whether to include a tab for the model coefficients

include_calib_summary
whether to include a tab for the model summary

include_calib_range
whether to include a tab for the evaluation range (if available, i.e. generated by iso_evaluate_calibration_range)

with_explicit_units
whether to include units in all column names

Value

returns the calib object invisibly for use in pipelines