Retrieve calibration data

Pull out data from the all_data nested calibration data frame (by default unnests all columnes). Typically used at various points after a dataset is prepared for calibration with iso_prepare_for_calibration, calibrations are generated with iso_generate_calibration and/or calibrations are applied with iso_apply_calibration. Note that when unnesting multi-row data (i.e. almost always), remaining data and calibration parameters are replicated across the unnested data set. This is NOT a problem for any downstream functions and in fact allows them to access the underlying calibrations if needed (e.g. in iso_mark_calibration_range). However, if this is undesired, please set keep_remaining_data = FALSE and keep_calibrations = FALSE when calling iso_get_calibration_data or use the iso_remove_list_columns function afterwards, which removes all nested information.

iso_get_calibration_data(
  dt,
  select = everything(),
  keep_remaining_data = keep_remaining_nested_data,
  keep_calibration_parameters = keep_other_list_data,
  keep_calibration_regressions = FALSE,
  quiet = default(quiet),
  keep_remaining_nested_data = TRUE,
  keep_other_list_data = TRUE
)

iso_unnest_data(...)

Arguments

dt
nested data frame with column all_data (see iso_prepare_for_calibration) and any number of calibrations (see iso_generate_calibration)

select
which columns to unnest - use c(...) to select multiple, supports all select syntax including renaming columns. Includes all columns by default (i.e. unnests all information).

keep_remaining_data
whether to keep the remaining parts of the partially unnested all_data data set (irrelevant if select = everything()). By default, remainig data is kept.

keep_calibration_parameters
whether to keep the calibration parameter columns (coefficients, summary, range) in the resulting data frame. By default, calibration parameters are kept to facilitate downstream operations that may need these calibration information (such as iso_mark_calibration_range).

keep_calibration_regressions
whether to keep the actual calibration regression objects in the resulting data frame. By default, regression objects are NOT kept because they lead to large data objects and do not usually have additional use after iso_apply_calibration has been called. There is rarely a reason to change this parameter to TRUE.

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.

keep_remaining_nested_data
DEPRECATED - renamed to keep_remaining_data

keep_other_list_data
DEPRECATED - renamed to keep_calibrations

...
additional parameters to forward to the func function

Details

iso_unnest_data is deprecated in favor of the more specific iso_get_calibration_data and will be removed in a future version of isoprocessor.

See also