Get calibration summary

Retrieve summary information for a calibration generated by iso_generate_calibration.

iso_get_calibration_summary(
  dt,
  calibration = last_calibration(dt),
  select = everything(),
  keep_calibration_parameters = keep_other_list_data,
  keep_calibration_regressions = FALSE,
  keep_other_calibrations = FALSE,
  quiet = default(quiet),
  keep_remaining_nested_data = FALSE,
  keep_other_list_data = FALSE
)

iso_unnest_calibration_summary(...)

Arguments

dt
nested data frame with column all_data (see iso_prepare_for_calibration) and any number of calibrations (see iso_generate_calibration)

select
which columns to unnest - use c(...) to select multiple, supports all select syntax including renaming columns. Includes all columns by default (i.e. unnests all information).

keep_calibration_parameters
whether to keep the calibration parameter columns (coefficients, summary, range) in the resulting data frame. By default, calibration parameters are kept to facilitate downstream operations that may need these calibration information (such as iso_mark_calibration_range).

keep_calibration_regressions
whether to keep the actual calibration regression objects in the resulting data frame. By default, regression objects are NOT kept because they lead to large data objects and do not usually have additional use after iso_apply_calibration has been called. There is rarely a reason to change this parameter to TRUE.

keep_other_calibrations
whether columns that belong to other calibrations (besides the one specified by calibration) should be kept in the data table. By default, they are removed for clarity.

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.

keep_remaining_nested_data
DEPRECATED - this parameter was removed for clarity and does no longer have any effect

keep_other_list_data
DEPRECATED - renamed to keep_calibrations

...
additional parameters to forward to the func function

Details

iso_unnest_calibration_summary is deprecated in favor of the equivalent iso_get_calibration_summary to standardize function call names. Please start using iso_get_calibration_summary as iso_unnest_calibration_summary will be deprecated in a future version of isoprocessor.

See also