Fetch problematic calibrations

Fetch data sets that have problematic calibrations (not enough data or another error occured during calibration) for further inspection. This function is typically called after iso_generate_calibration to inspect problematic data sets. It requires the columns generated by iso_generate_calibration and the same calibration parameter used there.

iso_get_problematic_calibrations(
  dt,
  calibration = last_calibration(dt),
  select = everything(),
  quiet = default(quiet)
)

Arguments

dt
nested data table with column all_data (see iso_prepare_for_calibration)

calibration
an informative name for the calibration (could be e.g. "d13C" or "conc"). If provided, will be used as a prefix for the new columns generated by this function. This parameter is most useful if there are multiple variables in the data set that need to be calibrated (e.g. multiple delta values, concentration, etc.). If there is only a single variable to calibrate, the calibration parameter is completely optional and can just be left blank (the default).

select
which columns to select for display - use c(...) to select multiple, supports all select syntax including renaming columns.

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.

See also