Visualize calibration parameters

Convenience function to visualize the residuals from calibrations generated by iso_generate_calibration. Uses iso_plot_data internally and additional parameters (...) can be passed to iso_plot_data.

iso_plot_calibration_parameters(dt, x = calibration_model_name(),
  calibration = last_calibration(dt), select_from_summary = c(R2 =
  adj.r.squared, RMSD = sigma), color = signif, panel = term ~ .,
  panel_scales = "free", points = TRUE, ...)

Arguments

dt
data frame to plot data from

x
the column for the x-axis aesthetic. Can be a numeric, text or datetime column (text and datetime column labels will be automatically rotated by 90 degrees). For clarity of the plot, only single variables are allowed. Use named vector with c(new_x = x) to rename the x variable on the fly. By default will show units in the axis names if there are any. If a datetime column is provided for x, parameters date_breaks (example: date_breaks = "2 hours") and date_labels can be set to fine-tune the x-axis appearance. See scale_date for additional details. Note that x can also be x = NULL for single data point plots - in this case the x axis is completely omitted.

select_from_summary
which parameters from the fit summary to include, by default includes the adjusted R2 (renamed just R2) and the root mean square deviation (RMSD), which R often calls sigma or residual standard deviation (often also called residual standard error and root mean square error instead of deviation, or standard error of the regression).

color
variable to use for color aesthetic for the plot

panel
whether to panel the data by anything. If using a single parameter (e.g. panel = panel), will generate a facet_wrap. If using a formula (e.g. panel = panel ~ . or panel = file_id ~ panel), will generate a facet_grid. The default for this parameter is to panel via facet grid by the y variable name but only if multiple y columns are provided. Otherwise will not generate any facets. If additional facet parameters are desired, please leave use facet_wrap and facet_grid diretly.

panel_scales
the scales parameter for the facets (if any are used)

points
whether to plot points (FALSE by default)

...
additional ggplot objects (e.g. geom_smooth()) that should be added to the plot PRIOR to the automatically generated layers for error bars, points and lines. Notet that to add geoms on top, please use regular iso_plot_data() + geom_smooth() + ... syntax instead.