Visualize the data
Source: R/visualization.R

General purpose convenience visualization function. Simply add other ggplot components after calling this function to customize more (e.g. with facet_wrap or theme calls).

iso_plot_data(dt, x, y, y_error = NULL, group = NULL, color = NULL,
  shape = NULL, size = NULL, linetype = NULL, lines = TRUE,
  points = FALSE)

Arguments

dt
data frame with peak data

x
the x axis aesthetic for the calibration parameters, can be datetime, text or numeric

y
which columns to visualize, combine with c()

y_error
an error column for drawing y error bars - if multiple y are provided, error needs to point to the same number of columns

group
what to group by, multiple columns allowed (combine with paste(...)), usually not necessary if groupings are fully defined through other aesthetics

color
variable to use for color aesthetic for the plot

shape
variable to use for shape aesthetic for the plot

size
variable to use for size aesthetic for the plot

linetype
variable to use for linetype aesthetic for the plot

lines
whether to plot lines (TRUE by default)

points
whether to plot points (FALSE by default)

Note

should probably make sure that the default columns for gather 'panel' and 'value' do not exist...

it would be great to allow renaming of the columns via this (especially the y column)