Plot raw data from isoreader files

Convenience function for making quick standard plots for raw isoreader data. Calls iso_plot_continuous_flow_data, iso_plot_dual_inlet_data and iso_plot_scan_data for data specific plotting (see those functions for parameter details). For customizing plotting calls, it is recommended to use iso_plot_continuous_flow_data, iso_plot_dual_inlet_data and iso_plot_scan_data directly.

iso_plot_raw_data(iso_files, ..., quiet = default(quiet))

Arguments

iso_files
collection of iso_file objects

...
parameters for the data specific plotting functions

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.

See also