Plot reference peaks

Visualize how consistent the reference peaks are across a serious of samples.

iso_plot_ref_peaks(
  dt,
  x,
  ratio,
  ...,
  group_id = file_id,
  is_ref_condition = TRUE,
  within_group = TRUE,
  is_ref_used = NULL
)

Arguments

dt
data table

x
which column to use for the x-axis

ratio
which ratio column(s) to compare for the reference peaks (can be multiple)

...
additional parameters passed to iso_plot_data

group_id
group identifier column(s) to clarify across which data groups the reference peak deviation should be calcualted. By default calculates reference peak variations within each analysis.

is_ref_condition
condition to identify which of the peaks are reference peaks (unless the peaks are prefilterd already). Must be a column or expression that evaluates to a logical (TRUE/FALSE).

within_group
deprectated, use group_id to group accordingly

is_ref_used
deprecated, set an aesthetics directly via ... paramter to iso_plot_data

See also