Set peak table from vendor data

Set the peak table from vendor data in each iso_file. Only the vendor data table columns that do exist in a file will be used, so this function can set peak table information across a variety of data files without errors. Provides a detailed info message about the columns that are used.

iso_set_peak_table_from_vendor_data_table(iso_files, direct_rename = c(),
  regex_rename = c(), quiet = default(quiet))

iso_set_peak_table_from_isodat_vendor_data_table(iso_files,
  quiet = default(quiet))

Arguments

direct_rename
select which columns to select and rename directly from the vendor data table. Must be a named vector with the values the vendor data table column names and the names the corresponding peak table column names.

regex_rename
select columns by regular expression and rename using str_replace. Must be a named vector with the values the regular expression to find vendor data table columns and the names the replacement expression for the correponding peak table column names.

Details

iso_set_peak_table_from_isodat_vendor_data_table provides specialized functionality to set peak table information from an Isodat vendor data tables. For compatibility with all downstream isoprocessor calculations, the resulting peak table has a very specific set of columns which are listed below. Mapping for Isodat data tables:

peak_nr: "Nr."

is_ref: "Is Ref.?"

rt_start: "Start"

rt: "Rt"

rt_end: "End"

ampX: "Ampl X"

bgrdX_start: "BGD X"

bgrdX_end: "BGD X" (Isodat only reports a single background amplitude)

areaX: "rIntensity X" (recorded intensities)

rX/Y: "rR X/Y" (recorded ratios)

rX/Y_ref: extrapolated reference ratio at the peak (not available from Isodat peak table)

rdX/Y: "rd X/Y" (recorded delta -> rX/Y / rX/Y_ref - 1, not shifted for true isotopic value of the ref)

dX/Y: "d X/Y" (delta frame shifted wrt to the true isotopic value of the ref; note that if this is a single element delta keeps just the numerator with the element symbol to fit conventional naming, e.g. d13C or d18O; whereas for a molecule omits the formula and keeps the mass ratios, e.g. d45/44)

atX: "AT

See also