Summarize peaks

Summarize peaks after peak mapping. This function is called after iso_map_peaks and can be used in combination with iso_get_problematic_peaks to inspect problematic peaks in particular. For the file_id parameter, make sure to use the same set or a subset of the columns used to identify individual files in the iso_map_peaks call before.

iso_summarize_peak_mappings(...)

# S3 method for iso_file_list
iso_summarize_peak_mappings(
  iso_files,
  include_file_info = NULL,
  compound = default(compound),
  rt = default(rt)
)

# S3 method for data.frame
iso_summarize_peak_mappings(
  peak_table,
  file_id = default(file_id),
  compound = default(compound),
  rt = default(rt)
)

Arguments

...
S3 method placeholder parameters, see class specific functions for details on parameters

iso_files
collection of continuous flow iso_file objects

include_file_info
which file information to include (see iso_get_file_info). Use c(...) to select multiple, supports all select syntax including renaming columns.

compound
the column in peak_maps that holds compound information

rt
the column in peak_table and colum prefix in peak_maps ("rt:...") that holds retention time information

peak_table
data frame with the peak table

file_id
the column(s) in peak_table that uniquely identify a file/set of peaks that belong together

Value

summary data table with one row for each unique combination of the file_id parameter

mapped: number of peaks that were identified during the mapping process (out of the total number of peaks in each sample)

ambiguous: number of mapped peaks that were ambiguous (out of all the mapped peaks) because they either have multiple matches or because they overlap with other mapped peaks

missing: number of peaks that were listed in the peak map but appear to be missing in the sample (out of the total number of peaks listed in the peak map)

peak_info: concatenated text with all the peaks and their retention times (with '?' for unknown peak names, retention times or other ambiguities

See also