Add metadata to data table

This function adds metadata flexibly by providing the possibility to match_by multiple different columns (paramter match_by) in sequence. This is equivalent to applying a set of increasingly more specific matching rules. For each metadata column, only the rows that have a defined non empty ("") value for that column will be mapped to the data. Each set of metadata can overwrite the previous matches such that the last metadata column defined by match_by will overwrite all previous matches for which it applies, even if they have already been a match for a previous column.

iso_add_metadata(dt, metadata, match_by = default(match_by),
  quiet = default(quiet))

Arguments

dt
data frame with the data

metadata
data frame with the metadata

match_by
the column (or columns) to match the metadata by. Used sequently, i.e. if the first column is defined in the metadata, it will be used first before mapping the remainder of the metadata with the second column, etc. All columns must exist in both the dt and metadata data frames.

Value

merged data frame with data and metadata and new column logical (TRUE/FALSE) has_metadata that holds information about which rows had metadata matches.

Details

This function also introduces a has_metadata column that would be typically used afterwards to inspect and/or filter data that has/doesn't have metadata. For routine downstream data processing, this function is thus usually followed by iso_get_missing_metadata to inspect which entries are missing metadata and iso_remove_missing_metadata to proceed only with entries that do have metadata assigned.

Note that this is a convenience function for easily adding metadata in a rule based way. If a direct join is suitable (i.e. if there is direct 1-to-1 or 1-to-many mapping of a single ID column or several id columns in combination), it will be a lot faster to use the join.

See also