Convert time units in continuous flow files

This function can be used to convert the time units of a collection of iso_files to a new common time unit. The original time units are inferred from the naming of the time column. New time units are very flexible and can be anything that duration understands, i.e. "s", "seconds", "min", "minutes", "hours", "days", etc. are all valid units.

iso_convert_time(iso_files, to, quiet = default(quiet))

Arguments

iso_files
collection of iso_file objects

to
what time units to convert to

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.

Value

the passed in iso_file(s) with changed time units