Visualize calibration parameters
Source: R/visualization.R

Visualize calibration coefficients and summary.

iso_plot_calibration_parameters(dt, calibration = "", x,
  select_from_summary = c(R2 = adj.r.squared, RMSD = sigma), color = NULL,
  shape = NULL, size = NULL, date_breaks = "2 hours",
  date_labels = "%d %b %H:%M")

Arguments

x
the x axis aesthetic for the calibration parameters, can be datetime, text or numeric

select_from_summary
which parameters from the fit summary to include, by default includes the adjusted R2 (renamed just R2) and the root mean square deviation (RMSD), which R often calls sigma or residual standard deviation (often also called residual standard error and root mean square error instead of deviation, or standard error of the regression).

color
variable to use for color aesthetic for the plot

shape
variable to use for shape aesthetic for the plot

size
variable to use for size aesthetic for the plot

date_breaks
what breaks to use for the x axis if it is a datetime

date_labels
datetime label pattern for x axis if it is a datetime